Add the top-level ES6 files
These were the easiest to implement, and look a lot like the original
CoffeeScript files.

Actually, that's why I went with [StandardJS] for the stylistic choice:
it can look like the CoffeeScript, so the maintainers can easily
transition to ESNext.

I also went with the 'not-so-standard' `es6` file extension to show the
difference between pre- and post-`babel` code.  If desired, they can all
safely be renamed to `js` with no compilation issues.

[StandardJS]: https://standardjs.com/