Convert the test suite to ES6 modules
The process for actually converting the test suite was actually
relatively painless.  It started with running coffeescript in-place,
running [`lebab`][lebab] against the resultant files with a few
different transforms (in particular the `multi-var`, `arrow`, and `let`
ones), renaming the files to `.es6` (again, an aesthetic choice) and
then going in and cleaning up some code.

In the helpers file, I just exported everything that had been declared
`global`.  It's a lot cleaner, and that stuff keeps its own context that
way.  After that, it was basically a matter of importing the new
exports.

I also changed the `for..in` boilerplate to a `map`, I think every time.
Yes, the boilerplate works, but `map`s, `forEach` and `for..of`
statements are a lot cleaner moving forward.

Up next: getting the tests to pass in ESNext-land.

[lebab]: https://www.npmjs.com/package/lebab