Update the test system to work as intended
I didn't realize that I had a typo in the name of the helper file being
included until I went to run the tests.

A few other changes happened, too: I'm pretty sure we can remove
`native-or-bluebird` in favor of `any-promise`, as the deprecation
warning states.  The original functions for some of the npm scripts now
have the ":coffee" suffix as they get transitioned away from, as well.

Lastly, the [`rimraf`][rimraf] and [`cross-env`][cross-env] packages
were introduced as devDependencies so we're not relying on POSIX
conventions to be there, even though my devbox is Antergos, for example.
The node version will have to be bumped up soon, but I don't know to
what version; I'm working in v7.8.0, personally.