Provide the minimum environment for the test suite
Gee, this was a difficult patch to implement.  However, the upside is
all the tests are failing *properly*.  The downside is that they're
*failing*.  That can be taken care of, though; especially since we can
now do proper TDD again.

In terms of new stuff, I updated all the outdated dependencies and
removed `native-or-bluebird` entirely in favor of `any-promise`.  The
old deps are still in the .coffee files, though, as I have been treating
them as read-only, for the most part.

I also added references to the eslint plugin [`import`][import], as well
as the babel plugin [`module-resolver`][module-resolver].  This is
because I was having trouble troubleshooting the cryptic `TypeError:
Cannot convert undefined or null to object` in the drops test file.

Lastly, as much as I like the `es6` extension, personally, none of the
*tools* (except `babel`) seem to feel the same way.  So I've renamed all
the files to `js` instead of using the extra configurations.  It'll make
it easier on other developers that way.

[import]: https://www.npmjs.com/package/eslint-plugin-import
[module-resolver]: https://www.npmjs.com/package/babel-plugin-module-resolver

Signed-off-by: Jonathan Sifuentes <jayands@users.noreply.github.com>