Update the test environment
[flowjs]: https://flowjs.org/

This patch moves Mocha's default options into a separate file,
making it easier to clean up the script's execution and be manually
overrideable.  It also adds a lot of the setup necessary to introduce
[Flow][flowjs], a tool for static analysis of JavaScript maintained by
facebook.

Up next: introducing Flow into the environment, and trying to fix
`StandardFilters` as it appears in `Engine`.

Signed-off-by: Jonathan Sifuentes <jayands@users.noreply.github.com>