Introduce Flow into the project
Changed the way some stuff is imported.  The most notable is that I couldn't
"export function default" in the StandardFilters set, because default is a
special keyword for exporting, but it *does* export an object named `default`
when you `export default X`.  So the way to import StandardFilters is `import *
as StandardFilters`, after which you can access StandardFilters.default for the
function/filter named `default`.

Also added Flow types to some of the files,
thereby starting the ability to do some static analysis.
I'm having problems with it in reference to generics,
though, but I think I figured it out.  I wanted Engine to take a generic type
`T` as promises only, so I tried to write it as `Engine<Promise<T>>`,
but that's not valid code, apparently.

Up next:

  [ ] Add more analysis annotations
    [ ] Get 20% annotated source code (by file count)
    [ ] Get 30% annotated source code (by file count)
    [ ] Get 40% annotated source code (by file count)
    [ ] Get 50% annotated source code (by file count)
    [ ] Get 60% annotated source code (by file count)
    [ ] Get 70% annotated source code (by file count)
    [ ] Get 80% annotated source code (by file count)
    [ ] Get 90% annotated source code (by file count)
    [ ] Get 100% annotated source code (by file count)
  [ ] Get all tests working
    [ ] Finish Filters
    [ ] Fix Engine's regiserFilters
    [ ] Figure out what is needed after, since most errors seem to be stemming
        from `Engine.registerFilters`

Signed-off-by: Jonathan Sifuentes <jayands@users.noreply.github.com>