Enable an apparently working file system
I'm not sure *where* exactly to find the error, but the reason three of
the tests are failing for `test/file_system.js` seems to be that the
error messages are *too* detailed for the assertion message of just
"error".

That said, everything *else* about them is correct, and I suspect the
errors are, too.

Signed-off-by: Jonathan Sifuentes <jayands@users.noreply.github.com>