Make the code "ready" for WIP commit
There seems to be a circular dependency, in Document -> Block ->Template
-> Document.  For some reason, it's only problematic in this form.

I actually did a local fork of the repo to try out a TypeScript version,
and while there are glaring static errors *there* still, this circular
dependency is not one of them.

Still, there should be enough here to get everyone else involved, so I'm
upstreaming this as a WIP commit.

Signed-off-by: Jonathan Sifuentes <jayands@users.noreply.github.com>